fix(driver): Apple Sign-In account matching + job feed updates
- AuthModels: omit userIdentifier when an identityToken is present so
  the backend derives the Apple user id from the verified JWT 'sub'
  claim — sending it explicitly bypassed JWT validation and matched
  the wrong (existing customer) account.
- JobFeedView: expanded feed UI; RatingView/JobCardView updates.
- Refresh driver app icons; untrack xcuserdata.

+++ JunkOS-Driver/Views/ActiveJob/RatingView.swift
@@ -166,7 +166,7 @@ struct RatingView: View {
                     Image(systemName: star <= viewModel.stars ? "star.fill" : "star")
                         .font(.system(size: 44))
                         .foregroundColor(star <= viewModel.stars ? .driverPrimary : .driverBorder)
-                        .scaleEffect(star <= viewModel.stars ? 1.1 : 1.0)
+                        .scaleEffect(star <= viewModel.stars ? CGFloat(1.1) : CGFloat(1.0))
                         .animation(AnimationConstants.bouncySpring, value: viewModel.stars)
                 }
                 .buttonStyle(.plain)
